+++ src/core/currency/wallet/currency-wallet-api.ts
@@ -640,13 +640,23 @@ export function makeCurrencyWalletApi(
       )
     },
     async parseUri(uri: string, currencyCode?: string): Promise<EdgeParsedUri> {
-      return await tools.parseUri(
+      const parsedUri = await tools.parseUri(
         uri,
         currencyCode,
         makeMetaTokens(
           input.props.state.accounts[accountId].customTokens[pluginId]
         )
       )
+
+      if (parsedUri.tokenId === undefined) {
+        const { tokenId = null } = upgradeCurrencyCode({
+          allTokens: input.props.state.accounts[accountId].allTokens[pluginId],
+          currencyInfo: plugin.currencyInfo,
+          currencyCode: parsedUri.currencyCode ?? currencyCode
+        })
+        parsedUri.tokenId = tokenId
+      }
+      return parsedUri
     },
 
     // Generic: